add updates

+++ backend/controllers/carController.js
@@ -1,5 +1,6 @@
 const Car = require('../models/carModel');
 // Controller valdo logika, kaip reaguoti i API uzklausas/requestus ir kreipiasi i Model, jeigu atitinka business logika (business logika - kad grazins visada)
+const Reservation = require('../models/reservationModel');
 
 exports.getCars = async (req, res) => {
   try {
@@ -11,15 +12,16 @@ exports.getCars = async (req, res) => {
 };
 
 exports.getCarById = async (req, res) => {
+  const carId = req.params.id;
   try {
-    const carId = req.params.id;
     const carById = await Car.findById(carId);
 
     if (!carById) {
       res.status(404).json({ error: 'Car not found' });
+      return;
     }
 
-    res.status(201).json({ message: 'Car found by id', carById });
+    res.status(200).json(carById);
   } catch (error) {
     res.status(500).json({ error: 'Server (controller getCarById) error' });
   }
@@ -41,6 +43,11 @@ exports.getCarById = async (req, res) => {
 // Update
 exports.updateCar = async (req, res) => {
   try {
+    if (req.user.role !== 'admin') {
+      return res
+        .status(403)
+        .json({ error: 'Not authorized. Admin access required' });
+    }
     const carId = req.params.id;
     const updatedCar = await Car.findByIdAndUpdate(carId, req.body, {
       new: true,
@@ -54,48 +61,43 @@ exports.updateCar = async (req, res) => {
   } catch (error) {
     res.status(500).json({ error: 'Server error' });
   }
-
-  // const carId = req.params.id;
-  // const car = Car.getCarById(carId);
-
-  // if (!car) {
-  //   return res.status(404).json({ message: 'Car not found' });
-  // }
-
-  // const updatedCar = Car.updateCar(carId, req.body);
-  // res.json(updatedCar);
 };
 
 // create new
-
+// ADMIN ONLY
 exports.addNewCar = async (req, res) => {
   try {
-    const newCar = await Car.create(req.body);
+    if (req.user.role !== 'admin') {
+      return res
+        .status(403)
+        .json({ error: 'Not authorized. Admin access required' });
+    }
+
+    const newCar = new Car(req.body);
     await newCar.save();
-    res.status(201).json({ message: 'Car created' });
+
+    res.status(201).json({ message: 'Car created successfully' });
   } catch (error) {
-    res.status(500).json({ error: 'Server error' });
+    res.status(500).json({
+      message: 'Failed to create car',
+    });
   }
-
-  // const { make, model } = req.body;
-  // const cars = Car.getAllCars();
-  // const newCar = {
-  //   id: cars.length + 1,
-  //   make: make,
-  //   model: model,
-  // };
-  // cars.push(newCar);
-
-  // // fs.writeFileSync(filePath, JSON.stringify(cars, null, 2));
-
-  // res.status(201).json(newCar);
 };
 
 //delete
 
 exports.deleteCar = async (req, res) => {
   try {
+    if (req.user.role !== 'admin') {
+      return res
+        .status(403)
+        .json({ error: 'Not authorized. Admin access required' });
+    }
     const carId = req.params.id;
+
+    // rezervaciju lenteleje suras su carId visas rezervacijas. Istrins visas kolekcijas su siuo id
+    await Reservation.deleteMany({ carId });
+
     const deletedCar = await Car.findByIdAndDelete(carId);
 
     if (!deletedCar) {
@@ -106,21 +108,4 @@ exports.deleteCar = async (req, res) => {
   } catch (error) {
     res.status(500).json({ error: 'Server error' });
   }
-  // const carId = req.params.id;
-  // const car = Car.getCarById(carId);
-
-  // if (!car) {
-  //   return res.status(404).json({ message: 'Car not found' });
-  // }
-
-  // Car.deleteCar(carId);
-  // res.status(204).send();
 };
-
-// module.exports = {
-//   getCars,
-//   getCarById,
-//   updateCar,
-//   addNewCar,
-//   deleteCar,
-// };
